requirements.txt
tiktoken
google-genai
openai
faiss-cpu
numpy